+++ CHANGELOG.md
@@ -13,6 +13,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
   [Samuel Giddins](https://github.com/segiddins)
   [#4716](https://github.com/CocoaPods/CocoaPods/issues/4716)
 
+* Don't warn about setting base config when identical to current config.  
+  [Jed Lewison](https://github.com/jedlewison)
+  [#4722](https://github.com/CocoaPods/CocoaPods/issues/4722)
+
 
 ## 1.0.0.beta.1 (2015-12-30)